dashboard/requirements.txt
streamlit>=1.29.0
pyvis>=0.3.1
networkx>=3.1
plotly>=5.18.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
PyYAML>=5.4
# Ajustado para compatibilidade com spaCy 3.7.0 e openai 1.12.0
pydantic>=2.0.0,<3.0.0
typing-extensions>=4.5.0

# Dependências para processamento de linguagem natural (NLP)
spacy>=3.7.0
pt_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0-py3-none-any.whl
scispacy>=0.5.3